[IGDM] Allocate memory for #VE info
The hardware writes information about the #VE
at the location referenced by the #VE info address
in the vmcs. This commit allocates this #VE info area
and initializes it to zero.

The four bytes starting at offset 4 of the #VE info area must be zero,
otherwise the #VE isn't delivered.